requirements.txt
streamlit
torch
sentence-transformers
networkx
scikit-learn
nltk
numpy
PyPDF2
gtts
pytesseract
Pillow
python-docx
